requirements.txt
channels
channels-redis

daphne
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt

Pillow
psycopg2

python-decouple
redis

cloudinary
django-cloudinary-storage
